Package refactoring: grails.plugins.jms -> grails.plugin.jms
