+++ build.gradle
@@ -1,16 +1,16 @@
 buildscript {
-	repositories {
-		maven { url 'https://repo.grails.org/grails/core' }
-	}
-	dependencies {
-		classpath "org.grails:grails-gradle-plugin:$grailsVersion"
-		classpath 'org.asciidoctor:asciidoctor-gradle-plugin:1.5.3'
-	}
+    repositories {
+        maven { url 'https://repo.grails.org/grails/core' }
+    }
+    dependencies {
+        classpath "org.grails:grails-gradle-plugin:$grailsVersion"
+        classpath 'org.asciidoctor:asciidoctor-gradle-plugin:1.5.3'
+    }
 }
 
 plugins {
-	id 'io.spring.dependency-management' version '0.5.4.RELEASE'
-	id 'com.jfrog.bintray' version '1.2'
+    id 'io.spring.dependency-management' version '0.5.4.RELEASE'
+    id 'com.jfrog.bintray' version '1.2'
 }
 
 apply from: 'gradle/plugin.gradle'
