+++ grails-app/conf/logback.groovy
@@ -1,7 +1,6 @@
 import grails.util.BuildSettings
 import grails.util.Environment
 
-
 // See http://logback.qos.ch/manual/groovy.html for details on configuration
 appender('STDOUT', ConsoleAppender) {
     encoder(PatternLayoutEncoder) {
@@ -17,7 +16,6 @@ logger('grails.app.service', INFO)
 if (Environment.current == Environment.DEVELOPMENT) {
     def targetDir = BuildSettings.TARGET_DIR
     if (targetDir) {
-
         appender("FULL_STACKTRACE", FileAppender) {
 
             file = "${targetDir}/stacktrace.log"
