+++ grails-app/services/grails/plugin/jms/JmsService.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package grails.plugins.jms
+package grails.plugin.jms
 
 import grails.plugins.jms.listener.GrailsMessagePostProcessor
