+++ src/main/groovy/grails/plugin/jms/Queue.java
