+++ src/main/groovy/grails/plugin/jms/bean/JmsBeanDefinitionBuilder.groovy
