+++ src/main/groovy/grails/plugin/jms/bean/JmsBeanDefinitionsBuilder.groovy
