+++ src/main/groovy/grails/plugin/jms/bean/JmsListenerAdapterAbstractBeanDefinitionBuilder.groovy
