+++ src/main/groovy/grails/plugin/jms/bean/JmsMessageConverterBeanDefinitionBuilder.groovy
