+++ src/main/groovy/grails/plugin/jms/bean/JmsTemplateBeanDefinitionBuilder.groovy
