+++ src/main/groovy/grails/plugin/jms/bean/MapBasedBeanDefinitionBuilder.groovy
