+++ src/main/groovy/grails/plugin/jms/listener/ListenerConfig.groovy
