+++ src/main/groovy/grails/plugin/jms/listener/ListenerConfigFactory.groovy
