+++ src/main/groovy/grails/plugin/jms/listener/adapter/LoggingListenerAdapter.groovy
