+++ src/main/groovy/grails/plugin/jms/listener/adapter/PersistenceContextAwareListenerAdapter.groovy
