+++ src/test/groovy/grails/plugin/jms/JmsServiceAsyncExecutorSpec.groovy
@@ -1,5 +1,6 @@
-package grails.plugins.jms
+package grails.plugin.jms
 
+import grails.plugin.jms.JmsService
 import spock.lang.AutoCleanup
 import spock.lang.Specification
