+++ src/test/groovy/grails/plugin/jms/JmsServiceConfSpec.groovy
@@ -1,5 +1,6 @@
-package grails.plugins.jms
+package grails.plugin.jms
 
+import grails.plugin.jms.JmsService
 import grails.util.Holders
 import org.grails.config.PropertySourcesConfig
 import spock.lang.Specification
