+++ src/test/groovy/grails/plugin/jms/bean/JmsBeanDefinitionBuilderTests.groovy
