+++ src/test/groovy/grails/plugin/jms/bean/JmsBeanDefinitionsBuilderTests.groovy
