+++ src/test/groovy/grails/plugin/jms/bean/JmsListenerAdapterAbstractBeanDefinitionBuilderTests.groovy
