+++ src/test/groovy/grails/plugin/jms/bean/JmsListenerContainerAbstractBeanDefinitionBuilderTests.groovy
