+++ src/test/groovy/grails/plugin/jms/bean/MapBasedBeanDefinitionBuilderTests.groovy
