+++ src/test/groovy/grails/plugin/jms/listener/ListenerConfigTests.groovy
