+++ src/test/groovy/grails/plugin/jms/listener/ServiceInspectorSpec.groovy
